+++ wgengine/wgcfg/config.go
@@ -9,6 +9,11 @@ import (
 	"inet.af/netaddr"
 )
 
+// EndpointDiscoSuffix is appended to the hex representation of a peer's discovery key
+// and is then the sole wireguard endpoint for peers with a non-zero discovery key.
+// This form is then recognize by magicsock's CreateEndpoint.
+const EndpointDiscoSuffix = ".disco.tailscale:12345"
+
 // Config is a WireGuard configuration.
 // It only supports the set of things Tailscale uses.
 type Config struct {
